feat: 통합 테스트 시 Redis 테스트 컨테이너 실행 추가 #19
