+++ src/test/java/org/songeun/petdongne_server/address/infrastructure/repository/LegalAddressSearchRepositoryTest.java
@@ -12,7 +12,6 @@ import org.songeun.petdongne_server.global.search.OrderedTokens;
 import org.songeun.petdongne_server.global.search.Token;
 import org.songeun.petdongne_server.global.util.GeoHashUtil;
 import org.songeun.petdongne_server.testSupport.IntegrationTestSupport;
-import org.songeun.petdongne_server.testSupport.PostgresSQLIntegrationTestSupport;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
@@ -23,7 +22,7 @@ import java.util.Set;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class LegalAddressSearchRepositoryTest extends PostgresSQLIntegrationTestSupport {
+class LegalAddressSearchRepositoryTest extends IntegrationTestSupport {
 
     @Autowired
     private LegalAddressSearchRepository searchRepository;
