+++ src/test/java/org/songeun/petdongne_server/building/infrastructure/BuildingSearchRepositoryTest.java
@@ -7,22 +7,15 @@ import org.songeun.petdongne_server.building.fixture.BuildingFixtureFactory;
 import org.songeun.petdongne_server.building.infrastructure.dto.BuildingGeoHashSearchQueryResponseDto;
 import org.songeun.petdongne_server.building.infrastructure.repository.BuildingSearchRepository;
 import org.songeun.petdongne_server.testSupport.IntegrationTestSupport;
-import org.songeun.petdongne_server.testSupport.PostgresSQLIntegrationTestSupport;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Profile;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
-import static org.hibernate.testing.transaction.TransactionUtil.doInJPA;
 
-class BuildingSearchRepositoryTest extends PostgresSQLIntegrationTestSupport {
+class BuildingSearchRepositoryTest extends IntegrationTestSupport {
 
     @Autowired
     private BuildingSearchRepository searchRepository;
