+++ src/test/java/org/songeun/petdongne_server/testSupport/IntegrationTestSupport.java
@@ -2,9 +2,51 @@ package org.songeun.petdongne_server.testSupport;
 
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.utility.DockerImageName;
 
 @SpringBootTest
 @ActiveProfiles("test")
 public abstract class IntegrationTestSupport {
 
+    // PostgreSQL Container
+    public static final String POSTGIS_IMAGE_NAME = "postgis/postgis:16-3.4-alpine";
+
+    static final PostgreSQLContainer<?> POSTGRE_SQL_CONTAINER =
+            new PostgreSQLContainer<>(DockerImageName.parse(POSTGIS_IMAGE_NAME)
+                    .asCompatibleSubstituteFor("postgres"))
+                    .withDatabaseName("testdb")
+                    .withUsername("testuser")
+                    .withPassword("testpass")
+                    .withInitScript("init-postgres.sql");
+
+    // Redis Container
+    private static final String REDIS_DOCKER_IMAGE = "redis:7.0.11-alpine";
+    private static final int REDIS_PORT = 6379;
+
+    static final GenericContainer<?> REDIS_CONTAINER =
+            new GenericContainer<>(DockerImageName.parse(REDIS_DOCKER_IMAGE))
+                    .withExposedPorts(REDIS_PORT)
+                    .withReuse(true);
+
+    static {
+        POSTGRE_SQL_CONTAINER.start();
+        REDIS_CONTAINER.start();
+    }
+
+    @DynamicPropertySource
+    static void configureProperties(DynamicPropertyRegistry registry) {
+        // PostgreSQL properties
+        registry.add("spring.datasource.url", POSTGRE_SQL_CONTAINER::getJdbcUrl);
+        registry.add("spring.datasource.username", POSTGRE_SQL_CONTAINER::getUsername);
+        registry.add("spring.datasource.password", POSTGRE_SQL_CONTAINER::getPassword);
+        registry.add("spring.datasource.driver-class-name", () -> "org.postgresql.Driver");
+
+        // Redis properties
+        registry.add("spring.data.redis.host", REDIS_CONTAINER::getHost);
+        registry.add("spring.data.redis.port", () -> REDIS_CONTAINER.getMappedPort(REDIS_PORT).toString());
+    }
 }
