--- src/test/java/org/songeun/petdongne_server/testSupport/PostgresSQLIntegrationTestSupport.java
@@ -1,35 +1,0 @@
-package org.songeun.petdongne_server.testSupport;
-
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.utility.DockerImageName;
-
-@SpringBootTest
-@ActiveProfiles("test")
-public abstract class PostgresSQLIntegrationTestSupport {
-
-    static DockerImageName postgis = DockerImageName.parse("postgis/postgis:16-3.4-alpine")
-            .asCompatibleSubstituteFor("postgres");
-
-    static final PostgreSQLContainer<?> POSTGRE_SQL_CONTAINER = new PostgreSQLContainer<>(postgis)
-            .withDatabaseName("testdb")
-            .withUsername("testuser")
-            .withPassword("testpass")
-            .withInitScript("init-postgres.sql");
-
-    static {
-        POSTGRE_SQL_CONTAINER.start();
-    }
-
-    @DynamicPropertySource
-    static void configureProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", POSTGRE_SQL_CONTAINER::getJdbcUrl);
-        registry.add("spring.datasource.username", POSTGRE_SQL_CONTAINER::getUsername);
-        registry.add("spring.datasource.password", POSTGRE_SQL_CONTAINER::getPassword);
-        registry.add("spring.datasource.driver-class-name", () -> "org.postgresql.Driver");
-    }
-
-}
